VM creator script & preseed
I choosed to make a creator script with functions to increase readability.

The preseed file is a bit raw, I don't understand it all yet.

For now it creates a VM with the openssh package.

I am working on another script to copy the SSH key on the VM.

I will merge the tow scripts later.

Script works and look nice, some improvments are in order tho :

* the VmName, UserName and PassWord are pre-set

It should be better with some kind of user input : file or form